examples/requirements.txt
uxapi
pyyaml
python-dotenv